requirements.txt
google-adk>=1.27.1
google-genai>=1.67.0
google-cloud-storage>=3.9.0
google-cloud-aiplatform>=1.141.0
moviepy>=2.2.1
pydantic>=2.12.5
python-dotenv>=1.0.0
